refactor(cli): extract MCP commands to dedicated module and fix Sse compilation
