+++ llms.txt
@@ -85,8 +85,6 @@ Created by: [cargo-llms-txt](https://github.com/masinc/cargo-llms-txt)
 - pub struct CliArgs
 - pub enum OutputFormat
 - pub enum Commands
-- pub enum McpCommands
-- pub struct McpServeArgs
 - pub struct InjectArgs
 - pub enum VectorSource
 - pub struct ProbeArgs
@@ -272,16 +270,23 @@ Created by: [cargo-llms-txt](https://github.com/masinc/cargo-llms-txt)
 - pub fn resolve_git_local_path
 - pub fn ensure_git_local_dir
 
+### src/cli/mcp.rs
+
+- pub enum McpCommands
+- pub struct McpServeArgs
+
 ### src/cli/mod.rs
 
 - pub mod args
 - pub mod commands
 - pub mod error
 - pub mod git_local
+- pub mod mcp
 - pub use args::*
 - pub use commands::{run_associate, run_associations, run_completions, run_delete, run_disassociate, run_export, run_get, run_import, run_index_dir, run_index_jsonl, run_inject, run_metrics, run_path, run_probe, run_probe_filtered, run_probe_graph, run_query, run_stats, run_traverse, run_update, run_watch}
 - pub use error::{CliError, ExitCode, Result}
 - pub use git_local::{ensure_git_local_dir, resolve_git_local_path}
+- pub use mcp::*
 
 ### src/concept_builder.rs
